requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.11
dataclasses==0.8
google-api-core==2.4.0
google-auth==2.6.0
google-cloud-pubsub==2.9.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
kafka-python==2.0.2
libcst==0.4.1
mypy-extensions==0.4.3
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.8
